+++ truck-geometry/src/decorators/rbf_surface/mod.rs
@@ -87,6 +87,26 @@ macro_rules! impl_radius_1dim {
 impl_radius_1dim!(BSplineCurve<Point1>);
 impl_radius_1dim!(NurbsCurve<Vector2>);
 
+impl<T: RadiusFunction> RadiusFunction for &T {
+    #[inline(always)]
+    fn der_n(&self, n: usize, t: f64) -> f64 { (**self).der_n(n, t) }
+}
+
+/// Oriented and reversible
+pub trait InvertibleRadiusFunction: RadiusFunction {
+    /// Inverts `self`
+    fn inverse(&self) -> Self;
+    /// Returns the inverse.
+    fn invert(&mut self);
+}
+
+impl InvertibleRadiusFunction for f64 {
+    #[inline(always)]
+    fn inverse(&self) -> Self { *self }
+    #[inline(always)]
+    fn invert(&mut self) {}
+}
+
 /// Contact point of the rolling ball and surface
 #[derive(Clone, Copy, Debug)]
 pub struct ContactPoint {
@@ -220,35 +240,13 @@ where
     }
 }
 
-impl<C, S0, S1, R> SearchParameter<D2> for RbfSurface<C, S0, S1, R>
-where
-    C: ParametricCurve3D + SearchNearestParameter<D1, Point = Point3>,
-    S0: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
-    S1: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
-    R: RadiusFunction,
-{
-    type Point = Point3;
-    fn search_parameter<H: Into<SPHint2D>>(
-        &self,
-        point: Self::Point,
-        hint: H,
-        trials: usize,
-    ) -> Option<(f64, f64)> {
-        let curve_hint = match hint.into() {
-            SPHint2D::Parameter(_, v) => SPHint1D::Parameter(v),
-            SPHint2D::Range(_, (v0, v1)) => SPHint1D::Range(v0, v1),
-            SPHint2D::None => SPHint1D::None,
-        };
-        let edge_curve = &self.edge_curve;
-        let v = edge_curve.search_nearest_parameter(point, curve_hint, trials)?;
-        let cc = self.contact_circle(v)?;
-
-        let cp0 = cc.contact_point0.point - cc.center;
-        let cp = point - cc.center;
-        let u = cp.angle(cp0).0 / cc.angle.0;
-        match cp.magnitude2().near(&cp0.magnitude2()) {
-            true => Some((u, v)),
-            false => None,
+impl<C, S0, S1, R> From<RbfSurface<C, S0, S1, R>> for RbfSurface<Box<C>, Box<S0>, Box<S1>, R> {
+    fn from(value: RbfSurface<C, S0, S1, R>) -> Self {
+        Self {
+            edge_curve: Box::new(value.edge_curve),
+            surface0: Box::new(value.surface0),
+            surface1: Box::new(value.surface1),
+            radius: value.radius,
         }
     }
 }
@@ -291,6 +289,7 @@ where
             _ => cc.contact_point1.point,
         }
     }
+
     fn der(&self, t: f64) -> Self::Vector { self.der_n(1, t) }
     fn der2(&self, t: f64) -> Self::Vector { self.der_n(2, t) }
     #[inline]
@@ -306,6 +305,52 @@ where
     S1: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
     R: RadiusFunction,
 {
+}
+
+impl<C, S0, S1, R> Cut for RbfContactCurve<C, S0, S1, R>
+where
+    C: ParametricCurve3D + Cut,
+    S0: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
+    S1: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
+    R: RadiusFunction,
+{
+    fn cut(&mut self, t: f64) -> Self {
+        let edge_curve = self.surface.edge_curve.cut(t);
+        Self {
+            surface: RbfSurface {
+                edge_curve,
+                surface0: self.surface.surface0.clone(),
+                surface1: self.surface.surface1.clone(),
+                radius: self.surface.radius.clone(),
+            },
+            index: self.index,
+        }
+    }
+}
+
+impl<C, S, R> Invertible for RbfContactCurve<C, S, S, R>
+where
+    C: ParametricCurve3D + Invertible,
+    S: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
+    R: InvertibleRadiusFunction,
+{
+    fn inverse(&self) -> Self {
+        Self {
+            surface: RbfSurface {
+                edge_curve: self.surface.edge_curve.inverse(),
+                surface0: self.surface.surface1.clone(),
+                surface1: self.surface.surface0.clone(),
+                radius: self.surface.radius.inverse(),
+            },
+            index: 1 - self.index,
+        }
+    }
+    fn invert(&mut self) {
+        self.surface.edge_curve.invert();
+        std::mem::swap(&mut self.surface.surface0, &mut self.surface.surface1);
+        self.surface.radius.invert();
+        self.index = 1 - self.index;
+    }
 }
 
 impl<C, S0, S1, R> ParameterDivision1D for RbfContactCurve<C, S0, S1, R>
@@ -324,8 +369,12 @@ where
 impl<C, S0, S1, R> SearchParameter<D1> for RbfContactCurve<C, S0, S1, R>
 where
     C: ParametricCurve3D + SearchNearestParameter<D1, Point = Point3>,
-    S0: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
-    S1: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
+    S0: ParametricSurface3D
+        + SearchParameter<D2, Point = Point3>
+        + SearchNearestParameter<D2, Point = Point3>,
+    S1: ParametricSurface3D
+        + SearchParameter<D2, Point = Point3>
+        + SearchNearestParameter<D2, Point = Point3>,
     R: RadiusFunction,
 {
     type Point = Point3;
@@ -335,16 +384,61 @@ where
         hint: H,
         trials: usize,
     ) -> Option<f64> {
-        let edge_curve = &self.surface.edge_curve;
-        let t = edge_curve.search_nearest_parameter(point, hint, trials)?;
-        let cc = self.surface.contact_circle(t)?;
-        let q = match self.index {
-            0 => cc.contact_point0.point,
-            _ => cc.contact_point1.point,
+        match self.index {
+            0 => self
+                .surface
+                .search_contact_curve0_parameter(point, hint, trials, true),
+            _ => RbfSurface {
+                edge_curve: &self.surface.edge_curve,
+                surface0: &self.surface.surface1,
+                surface1: &self.surface.surface0,
+                radius: &self.surface.radius,
+            }
+            .search_contact_curve0_parameter(point, hint, trials, false),
+        }
+    }
+}
+
+impl<C, S0, S1, R> SearchNearestParameter<D1> for RbfContactCurve<C, S0, S1, R>
+where
+    C: ParametricCurve3D + BoundedCurve,
+    S0: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
+    S1: ParametricSurface3D + SearchParameter<D2, Point = Point3>,
+    R: RadiusFunction,
+{
+    type Point = Point3;
+    fn search_nearest_parameter<H: Into<<D1 as SPDimension>::Hint>>(
+        &self,
+        point: Self::Point,
+        hint: H,
+        trials: usize,
+    ) -> Option<<D1 as SPDimension>::Parameter> {
+        use truck_geotrait::algo;
+        let hint = match hint.into() {
+            SPHint1D::Parameter(hint) => hint,
+            SPHint1D::Range(x, y) => {
+                algo::curve::presearch(self, point, (x, y), PRESEARCH_DIVISION)
+            }
+            SPHint1D::None => {
+                algo::curve::presearch(self, point, self.range_tuple(), PRESEARCH_DIVISION)
+            }
         };
-        match point.near(&q) {
-            true => Some(t),
-            false => None,
+        algo::curve::search_nearest_parameter(self, point, hint, trials)
+    }
+}
+
+impl<C, S0, S1, R> From<RbfContactCurve<C, S0, S1, R>>
+    for RbfContactCurve<Box<C>, Box<S0>, Box<S1>, R>
+{
+    fn from(value: RbfContactCurve<C, S0, S1, R>) -> Self {
+        Self {
+            surface: RbfSurface {
+                edge_curve: value.surface.edge_curve.into(),
+                surface0: value.surface.surface0.into(),
+                surface1: value.surface.surface1.into(),
+                radius: value.surface.radius,
+            },
+            index: value.index,
         }
     }
 }